test: deprecated 06_USM_Oracle.test.js
